+++ cheminformatics/cdk.json
@@ -21,10 +21,7 @@
     "number_of_azs": 2,
     "ec2_volume_size": 100,
     "instance_type": "p3.2xlarge",
-    "cheminformatics_container": "public.ecr.aws/b9g4r0v3/cheminformatics_demo:0.1.2",
-    "megamolbart_container": "nvcr.io/nvidia/clara/megamolbart:0.1.2",
-    "megamolbart_init_container": "public.ecr.aws/b9g4r0v3/megamolbart_init:latest",
-    "megamolbart_model_url": "https://api.ngc.nvidia.com/v2/models/nvidia/clara/megamolbart/versions/0.1/zip",
+    "megamolbart_container": "nvcr.io/nvidia/clara/megamolbart_v0.2:0.2.0",
     "@aws-cdk/aws-apigateway:usagePlanKeyOrderInsensitiveId": true,
     "@aws-cdk/core:stackRelativeExports": true,
     "@aws-cdk/aws-rds:lowercaseDbIdentifier": true,
