+++ cheminformatics/requirements.txt
@@ -1,2 +1,2 @@
-aws-cdk-lib==2.8.0
+aws-cdk-lib==2.33
 constructs>=10.0.0,<11.0.0
